ai_backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI/ML libraries
torch==2.1.0
torchvision==0.16.0
scikit-learn==1.3.2
scipy==1.11.4
numpy==1.24.3
pandas==2.1.3
transformers==4.35.2
openai==1.3.7
plotly==5.17.0
librosa==0.10.1
soundfile==0.12.1
joblib==1.3.2
cryptography==41.0.7
PyJWT==2.8.0

# Image processing
opencv-python==4.8.1.78
Pillow==10.1.0
pydicom==2.4.3

# PDF processing
pdfplumber==0.10.3
tabula-py==2.8.2

# Data processing
joblib==1.3.2

# HTTP client
httpx==0.25.2

# Logging and utilities
python-json-logger==2.0.7

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: GPU support (uncomment if using CUDA)
# torch==2.1.0+cu118
# torchvision==0.16.0+cu118 